requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
passlib==1.7.4
Pillow==10.0.1
pip==25.1.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.0.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rfc3986==2.0.0
rsa==4.9.1
schema==0.7.7
scipy==1.15.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.23.2
watchfiles==1.0.5
websockets==15.0.1
